requirements.txt
streamlit==1.36.0
huggingface_hub>=0.25.0
python-dotenv==1.0.1
